Add alias filename for email attachemnt
Added a new $alias param to the SMTP attach method which will allow
to set the filename of the attachment to a different filename other
then the file system filename. This is useful in cases where the
file system filename is hashed and/or versioned in some way and the
attachment filename needs to be different.